refactor: restructure project into multiple files
Splits the monolithic main.cpp into multiple files
for better organization and maintainability.
Includes auth, booking, movie, ui, and utils.
Also includes header files.
